Move route serialization into its own file. Placeholder in serializers for common waypoints serializer.
